build: -Wall -Wextra -Werror sur toutes les cibles — zéro warning toléré (B1)
L'argent réel ne se code pas avec des avertissements ignorés. Fixes :
sign-compare dans computeMetrics (BackTester), variable atrLast inutilisée
(DayTradeStrategy), dangling-else ×2 (tests), format-truncation (buffer du
test anti-look-ahead), et remplacement des agrégats BotState incomplets par
un helper makeBotState / une construction champ par champ — supprime au
passage la fragilité « l'ordre des champs est figé par les init agrégat »
notée dans Models.hpp. Aucun changement de comportement : 510/510 verts.

+++ include/backtest/BackTester.hpp
@@ -345,9 +345,9 @@ public:
         }
 
         // Sharpe + Volatilité + Sortino (rendements journaliers annualisés)
-        if (equityCurve.size() > warmup + 1) {
+        if (equityCurve.size() > static_cast<size_t>(warmup) + 1) {
             std::vector<double> rets;
-            for (size_t i = warmup + 1; i < equityCurve.size(); ++i)
+            for (size_t i = static_cast<size_t>(warmup) + 1; i < equityCurve.size(); ++i)
                 if (equityCurve[i-1] > 0)
                     rets.push_back((equityCurve[i] - equityCurve[i-1]) / equityCurve[i-1]);
